requirements.txt
amqp==2.6.1
arrow==1.2.2
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
binaryornot==0.4.4
celery==4.4.7
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cookiecutter==1.7.3
coverage==6.3.2
Deprecated==1.2.13
Django==4.0.3
django-browser-reload==1.3.0
django-filter==21.1
django-tailwind==3.1.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.1
jinja2-time==0.2.0
kombu==4.6.11
MarkupSafe==2.1.1
packaging==21.3
poyo==0.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2022.1
redis==4.2.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.9
vine==1.3.0
wrapt==1.14.0
